+++ core/src/widget_children/single_child_impl.rs
@@ -41,7 +41,14 @@ impl<T: SingleChild> SingleChild for FatObj<T> {
       .into_widget()
   }
 
-  fn into_parent(self: Box<Self>) -> Widget<'static> { self.into_widget() }
+  fn into_parent(self: Box<Self>) -> Widget<'static> {
+    let this = *self;
+    if !this.has_class() {
+      this.into_widget()
+    } else {
+      panic!("A FatObj should not have a class attribute when acting as a single parent")
+    }
+  }
 }
 
 macro_rules! impl_single_child_methods_for_pipe {